# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This API Handle MELI - Sprint 1 - Grupo 2",
    "title": "MELI Bootcamp API - Sprint 1 - Grupo 2",
    "termsOfService": "https://developers.mercadolibre.com.ar/es_ar/terminos-y-condiciones",
    "contact": {
      "name": "API Support",
      "url": "https://developers.mercadolibre.com.ar/support"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "paths": {
...
    "/api/v1/localities": {
      "get": {
        "description": "Get the details of all localities on the database.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Localities"
        ],
        "summary": "List Localities",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/web_utils.response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/entities.Locality"
                      }
                    }
                  }
                }
              ]
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          }
        }
      },
      "post": {
        "description": "search for a locality and return the number of sellers.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Localities"
        ],
        "summary": "CountSellers",
        "parameters": [
          {
            "type": "string",
            "description": "ID of Locality to be searched",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/web_utils.response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dtos.GetNumberOfSellersResponseDTO"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          }
        }
      }
    },
    "/api/v1/localities/{id}": {
      "get": {
        "description": "Get the details of a Locality",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Localities"
        ],
        "summary": "Get Locality",
        "parameters": [
          {
            "type": "string",
            "description": "ID of Locality to be searched",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/entities.Locality"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          }
        }
      },
      "delete": {
        "description": "Delete Localities",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Localities"
        ],
        "summary": "Delete Locality",
        "parameters": [
          {
            "type": "string",
            "description": "ID of a Locality to be excluded",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          }
        }
      },
      "patch": {
        "description": "Update the details of a Locality",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Localities"
        ],
        "summary": "Update Locality",
        "parameters": [
          {
            "type": "string",
            "description": "ID of Locality to be updated",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Updated Locality details",
            "name": "Locality",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dtos.UpdateLocalityRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/entities.Locality"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/web_utils.errorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "dtos.GetNumberOfSellersResponseDTO": {
      "type": "object",
      "properties": {
        "locality_id": {
          "type": "integer"
        },
        "locality_name": {
          "type": "string"
        },
        "sellers_count": {
          "type": "integer"
        }
      }
    },
...
    "dtos.UpdateLocalityRequestDTO": {
      "type": "object",
      "properties": {
        "country_name": {
          "type": "string"
        },
        "locality_name": {
          "type": "string"
        },
        "province_name": {
          "type": "string"
        }
      }
    },
...
    "entities.Locality": {
      "type": "object",
      "required": [
        "country_name",
        "id",
        "locality_name",
        "province_name"
      ],
      "properties": {
        "country_name": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "locality_name": {
          "type": "string"
        },
        "province_name": {
          "type": "string"
        }
      }
    },
...
    "web_utils.errorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "web_utils.response": {
      "type": "object",
      "properties": {
        "data": {}
      }
    }
  }
}